% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sample_sizes.R
\name{calc_sample_sizes}
\alias{calc_sample_sizes}
\title{Calculate sample sizes}
\usage{
calc_sample_sizes(d)
}
\arguments{
\item{d}{A data.frame produced by the \code{compile_tac_data} function containing 'target_name' and 'ct_value' columns.}
}
\value{
data.frame
}
\description{
This function takes a compiled data.frame from the \code{compile_tac_data} and calculates basic sample sizes for all targets.
}
\examples{
\dontrun{

compiled_tac <- read.csv('/Users/tac/compiled_tac.csv')
calc_sample_sizes(compiled_tac)

}
}
