% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casadelalibertad.R
\docType{package}
\name{casadelalibertad-package}
\alias{casadelalibertad}
\alias{casadelalibertad-package}
\title{casadelalibertad: Traduce al Español los Datos de 'Freedom House'}
\description{
Provee una versión traducida de todos los datasets del paquete 'freedomhouse'. English: Provides a Spanish translated version of the datasets in the 'freedomhouse' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pachadotdev/freedomhouse}
  \item Report bugs at \url{https://github.com/pachadotdev/freedomhouse/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas \email{m.sepulveda@mail.utoronto.ca} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

}
\keyword{internal}
