% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_pb.R
\name{.init_pb}
\alias{.init_pb}
\title{Initialize a progress bar}
\usage{
.init_pb(x)
}
\arguments{
\item{x}{Integer giving length of the for loop}
}
\description{
A wrapper function of the 'progress' package to initialize a default progress bar for long for loops
}
\examples{
\dontrun{
     len <- 100
     pb <- .init_pb(len)
     for (i in 1:len) {
          pb$tick()
          Sys.sleep(0.01)
          }
}

}
\keyword{internal}
