% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_admin_data.R
\name{download_admin_data}
\alias{download_admin_data}
\title{Download administrative boundaries from GADM}
\usage{
download_admin_data(iso3, path_output)
}
\arguments{
\item{iso3}{A three-letter capitalized character string. Must follow the ISO-3166 Alpha-3 country code
standard (\url{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-3}). Can accept a vector of multiple ISO codes.}

\item{path_output}{A character string giving the file path of an output directory to save downloaded data.}
}
\description{
This function takes a single ISO country code, or vector of multiple ISO country codes, and downloads the corresponding
high resolution administrative boundary shapefiles from \url{https://gadm.org/} via
the \href{https://rdrr.io/github/rspatial/geodata/man/gadm.html}{\code{geodata::gadm}} function. The most detailed administrative
units are returned (level 4 for most countries).
}
\examples{
\dontrun{

download_admin_data(iso3 = c('FRA', 'ITA'), path_output = getwd())

}
}
