% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_wopr_data.R
\name{download_wopr_data}
\alias{download_wopr_data}
\title{Download population raster data from WorldPop}
\usage{
download_wopr_data(iso3, output_path)
}
\arguments{
\item{iso3}{A three-letter capitalized character string. Must follow the ISO-3166 Alpha-3 country code
standard (\url{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-3}). Can accept a vector of multiple ISO codes.}

\item{output_path}{A character string giving the file path of an output directory to save downloaded data.}
}
\description{
This function takes a single ISO country code, or vector of multiple ISO country codes, and downloads the appropriate
100m resolution population count raster data from the WorldPop REST API. Note that these data are spatial
disaggregations of census data using random forest models described in \href{https://www.tandfonline.com/doi/full/10.1080/20964471.2019.1625151}{Lloyd et al. 2019}
and available for manual download at \url{https://hub.worldpop.org/geodata/listing?id=29}. Downloaded data sets
are saved to the \code{output_path} directory in .tif format.
}
\examples{
\dontrun{

download_wopr_data(iso3 = 'FRA', output_path = getwd())

}
}
