\name{ES-package}
\alias{ES-package}
\alias{ES-package}
\docType{package}
\title{
Edge Selection for Undirected Graphs
}
\description{
Implementation of the Edge Selection Algorithm
}
\details{
\tabular{ll}{
Package: \tab ESpackage\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-06-13\cr
License: \tab CRAN \cr
}
}
\author{
Meng Hwee Victor Ong, Sanjay Chaudhuri
}
\references{
Edge Selection for Undirected Graphs
}
\keyword{ES}
\keyword{cv.ES}
\keyword{ESpredict}