% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freedomhouse.R
\docType{package}
\name{freedomhouse-package}
\alias{freedomhouse}
\alias{freedomhouse-package}
\title{freedomhouse: Version en Castellano de los Datos de 'Freedom House's Freedom in the World'}
\description{
Desde 1973, Freedom House evalúa la situación de los derechos políticos y las libertades civiles en todo el mundo. La utilizan regularmente responsables de la formulación de políticas, periodistas, académicos, activistas y muchos otros. Este paquete facilita el uso de los datos de Freedom House's Freedom in the World en R. English: Since 1973, Freedom House has assessed the condition of political rights and civil liberties around the world. It is used on a regular basis by policymakers, journalists, academics, activists, and many others. This package makes it easier to use the Freedom House's Freedom in the World data within R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pacha.dev/freedomhouse/}
  \item Report bugs at \url{https://github.com/pachadotdev/freedomhouse/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas \email{m.sepulveda@mail.utoronto.ca} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

}
\keyword{internal}
