% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hydro_data.R
\name{get_hydro_data}
\alias{get_hydro_data}
\title{Get hydrological data}
\usage{
get_hydro_data(lon, lat, save_data = FALSE)
}
\arguments{
\item{lon}{A numeric vector giving the longitude of the sampling sites in Decimal Degrees.}

\item{lat}{A numeric vector giving the latitude of the sampling sites in Decimal Degrees.}

\item{save_data}{Logical indicating whether to save the intermediate raster layers used to
calculate the hydrological variables at each point. Default is \code{FALSE} where the temp directory is deleted.}
}
\value{
data.frame
}
\description{
This function takes information of where and when a set of environmental samples were
collected and retrieves a suite of topographical and hydrological variables for each unique
location. The variables include: elevation, slope, aspect, Topographical Wetness Index (TWI),
flow accumulation, total flow accumulation within 500m, and distance to the nearest stream.
The DEM is acquired via \href{https://www.rdocumentation.org/packages/elevatr/versions/0.99.0/topics/get_elev_raster}{\code{elevatr::get_elev_raster}}
and the suite of variables are calculated using functions from the \href{https://cran.r-project.org/web/packages/whitebox/index.html}{'WhiteboxTools'}
R frontend.
}
\examples{
\dontrun{

get_hydro_data(lon = template_es_data$lon,
               lat = template_es_data$lat,
               save_data = FALSE)

}
}
