% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_local_population_data.R
\name{get_local_population_data}
\alias{get_local_population_data}
\title{Get population counts for a local area}
\usage{
get_local_population_data(lon, lat, type)
}
\arguments{
\item{lon}{A numeric vector giving the longitude of the sampling sites in Decimal Degrees.}

\item{lat}{A numeric vector giving the latitude of the sampling sites in Decimal Degrees.}

\item{type}{A character string giving the type of local population to calculate. Options
are: 'catchment', 'radius', and 'admin'.}
}
\value{
data.frame
}
\description{
Get population counts for a local area
}
\examples{
\dontrun{

download_wopr_data(iso3 = 'FRA', output_path = getwd())

}
}
