% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_topo_data.R
\name{get_topo_data}
\alias{get_topo_data}
\title{Get topographical data}
\usage{
get_topo_data(lon, lat, show_plots = FALSE)
}
\arguments{
\item{lon}{A numeric vector giving the longitude of the sampling sites in Decimal Degrees.}

\item{lat}{A numeric vector giving the latitude of the sampling sites in Decimal Degrees.}

\item{show_plots}{Logical indicating whether to plot the raster and point data.}
}
\value{
data.frame
}
\description{
This function takes information of where and when a set of environmental samples were
collected and retrieves a 100m elevation raster around those locations and then calculates
the slope and aspect of each point. The DEM is acquired via \code{elevatr::get_elev_raster} and
topographical variables are calculated using \code{raster::terrain}.
}
\examples{
\dontrun{

get_topo_data(lon = template_es_data$lon,
              lat = template_es_data$lat,
              show_plots = TRUE)

}
}
