\docType{data}
\name{items_puntajes_por_categoria}
\alias{items_puntajes_por_categoria}
\title{Puntajes Desagregados de la Libertad en el Mundo, 2012-2022}
\format{Un data.frame de 25 filas y 4 columnas
\describe{
\item{item}{Letra del ítem (A-G).}
\item{sub_item}{Letra y número del sub-ítem (A1-G4).}
\item{descripcion_item}{Descripción del sub-ítem.}
\item{descripcion_sub_item}{Descripción del sub-ítem.}
}}
\usage{items_puntajes_por_categoria}
\description{Proporciona una descripción completa del significado de cada ítem y sub-ítem en la tabla de puntajes por categoría. Por ejemplo, el ítem A corresponde a Derechos Políticos.}
\source{Adaptado de Freedom House}
\keyword{datasets}
