% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logmean.R
\name{logmean}
\alias{logmean}
\title{Calculate the mean in log space}
\usage{
logmean(x)
}
\arguments{
\item{x}{A vector containing numeric values}
}
\value{
Scalar
}
\description{
This function gives the mean taken in log-scale for a set of numeric values. Values less than
or equal to zero are ignored as NA. Best suited for values taken from a highly skewed distribution,
as Ct values often are.
}
\examples{
\dontrun{

logmean(c(24.3, 10.3, 40, NA, 0, -1, 0.05))

}
}
