\name{marks}
\alias{marks}
\docType{data}
\title{
Mathematics Marks
}
\description{
Mathematic Marks from ggm package
}
\usage{data(marks)}
\format{
  A data frame with 88 observations on the following 5 variables.
  \describe{
    \item{\code{mechanics}}{a numeric vector}
    \item{\code{vectors}}{a numeric vector}
    \item{\code{algebra}}{a numeric vector}
    \item{\code{analysis}}{a numeric vector}
    \item{\code{statistics}}{a numeric vector}
  }
}
\details{
Mechanics and Vectors were closed book examinations. Algebra, Analysis and Statistics were open book examinations.
}
\source{
Mardia, K.V., Kent, J.T. and Bibby, (1979). Multivariate analysis. London: Academic Press.
}
\references{
Whittaker, J. (1990). Graphical models in applied multivariate statistics. Chichester: Wiley.
}
\examples{
data(marks)
}
\keyword{datasets}
