% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_wopr_api.R
\name{query_wopr_api}
\alias{query_wopr_api}
\title{Query the WorldPop REST API}
\usage{
query_wopr_api(iso3)
}
\arguments{
\item{iso3}{A three-letter capitalized character string. Must follow the ISO-3166 Alpha-3 country code
standard (\url{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-3}).}
}
\value{
list
}
\description{
This function queries the WorldPop REST API for all population raster metadata associated with a particular ISO country code.
If no ISO country code is provided, the function returns metadata for all available ISO codes on the WorldPop API.
}
\examples{
\dontrun{

query_wopr_api('USA')
query_wopr_api(NULL)

}
}
