% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_es_data.R
\docType{data}
\name{template_es_data}
\alias{template_es_data}
\title{Template environmental sampling data}
\format{
\subsection{\code{template_es_data}}{

A data frame with 6 columns:
\describe{
\item{date}{The date each sample was collected. Formate is "YYYY-MM-DD".}
\item{location_id}{A unique identifier for each of the sampling locations.}
\item{lat}{The lattitude of the sampling location in decimal degrees.}
\item{lon}{The longitude of the sampling location in decimal degrees.}
\item{target_name}{The unique name of the gene target for which the Ct values correspond.}
\item{ct_value}{The Cycle Threshold (Ct) of the qPCR assay.}
}
}
}
\usage{
template_es_data
}
\description{
The \code{template_es_data} object provides a template of the data format required by the \code{es} package.
}
\keyword{datasets}
