% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_delta_delta_ct.R
\name{apply_delta_delta_ct}
\alias{apply_delta_delta_ct}
\title{Apply the delta delta Ct calculation to a data.frame}
\usage{
apply_delta_delta_ct(
  df,
  target_names,
  reference_names,
  amplification_efficiency = 1
)
}
\arguments{
\item{df}{A data.frame of class \code{esdata}.}

\item{target_names}{Character vector giving the names of the target genes.}

\item{reference_names}{Character vector giving the names of the reference genes associated with each target gene.}

\item{amplification_efficiency}{A scalar between 0 and 1 giving the assumed PCR amplification efficiency for all samples in the equation. Defaults to 1, which assumes 100\% efficiency.}
}
\value{
data.frame
}
\description{
This function will calculate the delta delta Ct metric for all applicable observations in a data.frame
by applying the \code{calc_delta_delta_ct} function. The data.frame must have the following columns:
'location_id', 'sample_date', 'target_name', and 'ct_value'. The relevant target_names and and associated reference_names
must be provided. The result is a data.frame containing a 'delta_delta_ct' column which can be merge into the source data.frame.
}
\examples{
\dontrun{

df_example <- data.frame(location_id = c(rep(1,9), rep(2,9), rep(3,9)),
                         sample_date = as.Date(rep(c(rep('2024-01-01',3), rep('2024-01-02',3), rep('2024-01-03',3)), 3)),
                         target_name = rep(c('gene1', 'gene2', 'housekeeping'), 9),
                         ct_value = as.numeric(runif(27, min=5, max=40)))

df_result <- apply_delta_delta_ct(df = df_example,
                                  target_names = c('gene1', 'gene2'),
                                  reference_names = c('housekeeping', 'housekeeping'),
                                  amplification_efficiency = 0.95)

merge(df_example, df_result, by=c('location_id', 'sample_date', 'target_name'), all.x=TRUE)

}
}
