% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sample_sizes.R
\name{calc_sample_sizes}
\alias{calc_sample_sizes}
\title{Calculate sample sizes}
\usage{
calc_sample_sizes(df, cutoff = 40)
}
\arguments{
\item{df}{A data.frame produced by the \code{compile_tac_data} function containing 'target_name' and 'ct_value' columns.}

\item{cutoff}{Numeric scalar giving the cutoff Ct value over which a gene target is deemed absent from a sample. Default is 40.}
}
\value{
data.frame
}
\description{
This function takes a compiled data.frame following the format shown in the \code{template_es_data} object
and calculates basic sample sizes and detection rates for all gene targets.
}
\examples{
\dontrun{

calc_sample_sizes(template_es_data)

}
}
