% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_admin_data.R
\name{get_admin_data}
\alias{get_admin_data}
\title{Get administrative data for a set of points}
\usage{
get_admin_data(lon, lat, path_admin_data)
}
\arguments{
\item{lon}{A numeric vector giving the longitude of the sampling sites in Decimal Degrees.}

\item{lat}{A numeric vector giving the latitude of the sampling sites in Decimal Degrees.}

\item{path_admin_data}{The file path to the admin data. Note that the function expects .rds format output from the \code{download_admin_data} function.}
}
\value{
data.frame
}
\description{
This function takes a set of longitude and latitude coordinates and retrieves the administrative
units that each point lies within. The high resolution shapefiles used to determine the administrative boundaries
are acquired from \url{https://gadm.org/} via the \href{https://rdrr.io/github/rspatial/geodata/man/gadm.html}{\code{geodata::gadm}} function.
}
\examples{
\dontrun{

download_admin_data(iso3 = "BGD", path_output = getwd())

get_admin_data(lon = template_es_data$lon,
               lat = template_es_data$lat,
               path_admin_data = file.path(getwd(), 'gadm/gadm41_BGD_4_pk.rds'))

}
}
